feat: rename path package to home
