+++ session/list.go
@@ -3,8 +3,8 @@ package session
 import (
 	"fmt"
 
+	"github.com/joshmedeski/sesh/home"
 	"github.com/joshmedeski/sesh/model"
-	"github.com/joshmedeski/sesh/path"
 	"github.com/joshmedeski/sesh/tmux"
 	"github.com/joshmedeski/sesh/zoxide"
 )
@@ -31,7 +31,7 @@ func (s *RealSession) List(opts ListOptions) ([]model.SeshSession, error) {
 	}
 
 	if srcs["zoxide"] {
-		zoxideList, err := listZoxideResults(s.zoxide, s.path)
+		zoxideList, err := listZoxideResults(s.zoxide, s.home)
 		if err != nil {
 			return nil, err
 		}
@@ -69,14 +69,14 @@ func listTmuxSessions(t tmux.Tmux) ([]model.SeshSession, error) {
 	return sessions, nil
 }
 
-func listZoxideResults(z zoxide.Zoxide, p path.Path) ([]model.SeshSession, error) {
+func listZoxideResults(z zoxide.Zoxide, h home.Home) ([]model.SeshSession, error) {
 	zoxideResults, err := z.ListResults()
 	if err != nil {
 		return nil, fmt.Errorf("couldn't list zoxide results: %q", err)
 	}
 	sessions := make([]model.SeshSession, len(zoxideResults))
 	for i, r := range zoxideResults {
-		name, err := p.ShortenHome(r.Path)
+		name, err := h.ShortenHome(r.Path)
 		if err != nil {
 			return nil, fmt.Errorf("couldn't shorten path: %q", err)
 		}
