+++ session/session.go
@@ -1,8 +1,8 @@
 package session
 
 import (
+	"github.com/joshmedeski/sesh/home"
 	"github.com/joshmedeski/sesh/model"
-	"github.com/joshmedeski/sesh/path"
 	"github.com/joshmedeski/sesh/tmux"
 	"github.com/joshmedeski/sesh/zoxide"
 )
@@ -12,11 +12,11 @@ type Session interface {
 }
 
 type RealSession struct {
-	path   path.Path
+	home   home.Home
 	tmux   tmux.Tmux
 	zoxide zoxide.Zoxide
 }
 
-func NewSession(path path.Path, tmux tmux.Tmux, zoxide zoxide.Zoxide) Session {
-	return &RealSession{path, tmux, zoxide}
+func NewSession(home home.Home, tmux tmux.Tmux, zoxide zoxide.Zoxide) Session {
+	return &RealSession{home, tmux, zoxide}
 }
